Refactor connector and component configurations for projects 1 and 2; update test cases and expected outputs
