+++ tests/src/test/resources/expected/project1/connector.xml
@@ -1,23 +1,27 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-  ~Copyright (c) 2025, WSO2 LLC. (https://www.wso2.com).
-  ~~WSO2 LLC. licenses this file to you under the Apache License,
-  ~Version 2.0 (the "License"); you may not use this file except
-  ~in compliance with the License.
-  ~You may obtain a copy of the License at
-  ~~http://www.apache.org/licenses/LICENSE-2.0
-  ~~Unless required by applicable law or agreed to in writing,
-  ~software distributed under the License is distributed on an
-  ~"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  ~KIND, either express or implied.  See the License for the
-  ~specific language governing permissions and limitations
-  ~under the License.
+  ~ Copyright (c) 2025, WSO2 LLC. (https://www.wso2.com).
+  ~
+  ~ WSO2 LLC. licenses this file to you under the Apache License,
+  ~ Version 2.0 (the "License"); you may not use this file except
+  ~ in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~ http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing,
+  ~ software distributed under the License is distributed on an
+  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+  ~ KIND, either express or implied.  See the License for the
+  ~ specific language governing permissions and limitations
+  ~ under the License.
   -->
 <connector>
     <component name="project1" package="io.ballerina.stdlib.mi">
-        <dependency component="test"/>
+        
+        <dependency component="functions"/>
         <description>helps to connect with external systems</description>
     </component>
     <icon>icon/icon-small.gif</icon>
-    <type>module</type>
+    <displayName>project1</displayName>
 </connector>
