+++ tests/src/test/resources/expected/project1/functions/test.xml
@@ -16,22 +16,25 @@
 <template xmlns="http://ws.apache.org/ns/synapse" name="test">
     <parameter name="responseVariable" description="The target variable in which the output of the operation will be stored"/>
     <parameter name="overwriteBody" description="Replace the Message Body in Message Context with the output payload of the operation"/>
-    <parameter name="xmlA" description="Component description"/>
-<!--    TODO: if union, the selected member type becomes a parameter-->
-    <parameter name="xmlB" description="Component description"/>
-<!--    TODO: if union, the selected member type becomes a parameter-->
-    <parameter name="xmlC" description="Component description"/>
-<!--    TODO: if union, the selected member type becomes a parameter-->
+    <!-- TODO: if union, the selected member type becomes a parameter-->
+    <parameter name="xmlA" description=""/>
+    <parameter name="xmlB" description=""/>
+    <parameter name="xmlC" description=""/>
     <sequence>
-        <property name="paramSize" value="3"/>
-        <property name="paramFunctionName" value="test"/>
         <property name="param0" value="xmlA"/>
         <property name="paramType0" value="xml"/>
         <property name="param1" value="xmlB"/>
         <property name="paramType1" value="xml"/>
         <property name="param2" value="xmlC"/>
         <property name="paramType2" value="xml"/>
+        <property name="paramSize" value="3"/>
+        <property name="paramFunctionName" value="test"/>
         <property name="returnType" value="xml"/>
-        <class name="io.ballerina.stdlib.mi.Mediator"/>
+
+        <class name="io.ballerina.stdlib.mi.Mediator">
+            <property name="orgName" value="org"/>
+            <property name="moduleName" value="project1"/>
+            <property name="version" value="0"/>
+        </class>
     </sequence>
 </template>
