--- tests/src/test/resources/expected/project1/outputschema/test.json
@@ -1,14 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "type": "object",
-  "title": "Output Schema for test operation",
-  "description": "Output schema for the test operation.",
-  "properties": {
-    "payload": {
-      "type": "string",
-      "description": "The output string from the operation."
-    }
-  },
-  "required": ["payload"],
-  "additionalProperties": false
-}
