+++ tests/src/test/resources/expected/project2/uischema/test.json
@@ -2,55 +2,20 @@
   "connectorName": "project2",
   "operationName": "test",
   "title": "test",
-  "help": "Input parameters for test operation",
+  "help": "",
   "elements": [
     {
       "type": "attributeGroup",
       "value": {
         "groupName": "General",
         "elements": [
+          
           {
             "type": "attributeGroup",
             "value": {
               "groupName": "Input Variables",
               "elements": [
-                
-                {
-                  "type": "attribute",
-                  "value": {
-                    "name": "xmlA",
-                    "displayName": "xmlA",
-                    "inputType": "stringOrExpression",
-                    "defaultValue": "",
-                    "required": "true",
-                    "helpTip": "Input parameters required to invoke the ballerina function"
-                  }
-                },
-                
-                {
-                  "type": "attribute",
-                  "value": {
-                    "name": "xmlB",
-                    "displayName": "xmlB",
-                    "inputType": "stringOrExpression",
-                    "defaultValue": "",
-                    "required": "true",
-                    "helpTip": "Input parameters required to invoke the ballerina function"
-                  }
-                },
-                
-                {
-                  "type": "attribute",
-                  "value": {
-                    "name": "xmlC",
-                    "displayName": "xmlC",
-                    "inputType": "stringOrExpression",
-                    "defaultValue": "",
-                    "required": "true",
-                    "helpTip": "Input parameters required to invoke the ballerina function"
-                  }
-                }
-                
+                ,,
               ]
             }
           },
@@ -66,7 +31,7 @@
                     "displayName": "Output Variable Name",
                     "inputType": "string",
                     "defaultValue": "",
-                    "required": "true",
+                    "required": "false",
                     "helpTip": "The target variable in which the output of the operation will be stored"
                   }
                 },
@@ -78,8 +43,7 @@
                     "inputType": "checkbox",
                     "defaultValue": "false",
                     "helpTip": "Replace the Message Body in Message Context with the output payload of the operation (This will remove the payload from the above variable. But other elements such as headers and attributes will remain).",
-                    "required": "false",
-                    "hidden": "true"
+                    "required": "false"
                   }
                 }
               ]
